feat: Update ShowFoodMenu component to include "+" button for each food item. Rollback changes to FoodMenu.jsx component.
